util: move string_in_list to util_ext
To be able to include string_in_list() without additional
dependencies it is moved into a separate file.

Related to https://pagure.io/SSSD/sssd/issue/3050
